menu_service/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
psycopg2_binary==2.9.7
dynaconf==3.2.2
djangorestframework-simplejwt==5.3.0